Make ruleset a type for brevity
