+++ internal/rules/gameoflife_test.go
